ShippingApp/_BACKUP_TO_DELETE/requirements.txt
flask
flask-cors
psycopg2-binary
fuzzywuzzy
python-Levenshtein
